requirement.txt
brew_distance==1.0.1
alignment==1.0.10
deap==1.3.1
nltk==3.5
pandas==0.25.2
joblib==0.15.1
regex==2020.6.8
numpy==1.18.5
scipy==1.4.1
scikit-learn==0.23.1
sklearn==0.0
keras==2.3.1
tensorflow==2.2.0
matplotlib==3.2.2
pathlib==1.0.1
pyopencl==2020.1
h5py==2.10.0
pendulum==2.1.0
python-dateutil==2.6.1
docutils==0.10
